Finish burning down kotterknife
