+++ app/src/main/assets/licenses_github.json
@@ -9,10 +9,6 @@
   },
   {
     "owner": "jakewharton",
-    "name": "kotterknife"
-  },
-  {
-    "owner": "jakewharton",
     "name": "processphoenix"
   },
   {
