+++ app/src/main/kotlin/io/sweers/catchup/ui/activity/ImageViewerActivity.kt
@@ -48,6 +48,7 @@ import coil.transform.Transformation
 import io.sweers.catchup.R
 import io.sweers.catchup.databinding.ActivityImageViewerBinding
 import io.sweers.catchup.ui.immersive.SystemUiHelper
+import io.sweers.catchup.util.setContentView
 import io.sweers.catchup.util.toggleVisibility
 import me.saket.flick.ContentSizeProvider2
 import me.saket.flick.FlickCallbacks
