+++ libraries/base-ui/src/main/kotlin/io/sweers/catchup/base/ui/BaseFragment.kt
@@ -61,7 +61,6 @@ abstract class BaseFragment<T : ViewBinding> : Fragment(), ScopeProvider, Backpr
 
   override fun onDestroyView() {
     dayOnlyContext = null
-    KotterKnife.reset(this)
     super.onDestroyView()
   }
